+++ src/components/ReportBalance/ReportBalance.jsx
@@ -1,5 +1,5 @@
 import { useSelector } from "react-redux";
-import { getExpenseUser } from "../../redux/getBalance/balance-selector";
+// import { getExpenseUser } from "../../redux/getBalance/balance-selector";
 import { getDataForInput } from "../../redux/setDataDiagram/dataDiagram-selector";
 import {
   ReportBalanceWrapper,
